+++ packages/bottom-sheet/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useVisibleViewportSize } from './use-visualviewport-size';