+++ packages/checkbox-group/src/Component.responsive.tsx
@@ -13,15 +13,21 @@ export type CheckboxGroupProps = Omit<BaseCheckboxGroupProps, 'styles'> & {
      * @default 1024
      */
     breakpoint?: number;
+
+    /**
+     * Значение по-умолчанию для хука useMatchMedia
+     */
+    defaultMatchMediaValue?: boolean | (() => boolean);
 };
 
 export const CheckboxGroup: FC<CheckboxGroupProps> = ({
     breakpoint = getComponentBreakpoint(),
+    defaultMatchMediaValue,
     ...restProps
 }) => {
     const query = `(min-width: ${breakpoint}px)`;
 
-    const [isDesktop] = useMatchMedia(query);
+    const [isDesktop] = useMatchMedia(query, defaultMatchMediaValue);
 
     const Component = isDesktop ? CheckboxGroupDesktop : CheckboxGroupMobile;