+++ packages/plate/src/Component.responsive.tsx
@@ -13,13 +13,21 @@ export type PlateProps = Omit<BasePlateProps, 'styles'> & {
      * @default 1024
      */
     breakpoint?: number;
+
+    /**
+     * Значение по-умолчанию для хука useMatchMedia
+     */
+    defaultMatchMediaValue?: boolean | (() => boolean);
 };
 
 export const Plate = forwardRef<HTMLDivElement, PlateProps>(
-    ({ children, breakpoint = getComponentBreakpoint(), ...restProps }, ref) => {
+    (
+        { children, breakpoint = getComponentBreakpoint(), defaultMatchMediaValue, ...restProps },
+        ref,
+    ) => {
         const query = `(min-width: ${breakpoint}px)`;
 
-        const [isDesktop] = useMatchMedia(query);
+        const [isDesktop] = useMatchMedia(query, defaultMatchMediaValue);
 
         const Component = isDesktop ? PlateDesktop : PlateMobile;