+++ packages/accordion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.1.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.7.0
+
 ## 1.1.0
 
 ### Minor Changes