+++ packages/gallery/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 5.6.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.7.0
+    -   tooltip@7.3.19
+    -   icon-button@6.7.3
+
 ## 5.6.3
 
 ### Patch Changes