+++ packages/input/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 15.0.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   mq@4.3.0
+    -   button@11.5.3
+    -   form-control@12.2.5
+
 ## 15.0.1
 
 ### Patch Changes