+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 9.3.8
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   mq@4.3.0
+    -   navigation-bar-private@0.1.7
+
 ## 9.3.7
 
 ### Patch Changes