+++ packages/slider-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 8.2.15
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@15.0.2
+
 ## 8.2.14
 
 ### Patch Changes