+++ packages/tab-bar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.0.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.7.0
+
 ## 2.0.4
 
 ### Patch Changes